Optional build configuration `derelict-gl-library`; uses Derelict for OpenGL
Derelict is the de-facto standard for GL in the D community and quite a
few of my projects depend on it. This changes makes using dimgui with
these projects a matter of selecting the right subconfiguration
inststead of rewriting a bunch of code. The optional build config simply
changes dependencies and uses a `version` to import Derelict instead of
glad-drey.